requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
WTForms==3.1.1

# Database
PyMySQL==1.1.0
SQLAlchemy==2.0.34

# Security and validation
Werkzeug==3.0.1
MarkupSafe==2.1.3
python-dotenv==1.0.0
Pillow>=10.0.0

# Google OAuth and APIs
google-auth>=2.17.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.100.0
requests>=2.31.0

# Email (fallback SMTP)
secure-smtplib>=0.1.1

# Optional: For enhanced security
cryptography>=41.0.0